requirements.txt
torch
matplotlib
gradio
numpy
pandas
lightning
pillow